+++ src/main/java/net/yxiao233/ifeu/common/compact/kubejs/ModKubeJSPlugin.java
@@ -6,7 +6,7 @@ import dev.latvian.mods.kubejs.registry.RegistryInfo;
 import net.yxiao233.ifeu.common.compact.kubejs.items.EfficiencyAddonItemBuilder;
 import net.yxiao233.ifeu.common.compact.kubejs.items.ProcessingAddonItemBuilder;
 import net.yxiao233.ifeu.common.compact.kubejs.items.SpeedAddonItemBuilder;
-import net.yxiao233.ifeu.common.compact.kubejs.schemas.InfuserSchema;
+import net.yxiao233.ifeu.common.compact.kubejs.schemas.*;
 
 public class ModKubeJSPlugin extends KubeJSPlugin {
 
@@ -20,6 +20,14 @@ public class ModKubeJSPlugin extends KubeJSPlugin {
     @Override
     public void registerRecipeSchemas(RegisterRecipeSchemasEvent event) {
         event.namespace("ifeu")
-                        .register("infuser", InfuserSchema.SCHEMA);
+                .register("infuser", InfuserSchema.SCHEMA);
+
+        event.namespace("industrialforegoing")
+                .register("crusher", CrusherSchema.SCHEMA)
+                .register("dissolution_chamber", DissolutionChamberSchema.SCHEMA)
+                .register("fluid_extractor", FluidExtractorSchema.SCHEMA)
+                .register("laser_drill_fluid", LaserDrillFluidSchema.SCHEMA)
+                .register("laser_drill_ore", LaserDrillOreSchema.SCHEMA)
+                .register("stonework_generate", StoneWorkGenerateSchema.SCHEMA);
     }
 }