+++ src/main/java/net/yxiao233/ifeu/common/compact/kubejs/schemas/CrusherSchema.java
@@ -1,0 +1,13 @@
+package net.yxiao233.ifeu.common.compact.kubejs.schemas;
+
+import dev.latvian.mods.kubejs.item.InputItem;
+import dev.latvian.mods.kubejs.item.OutputItem;
+import dev.latvian.mods.kubejs.recipe.RecipeKey;
+import dev.latvian.mods.kubejs.recipe.component.ItemComponents;
+import dev.latvian.mods.kubejs.recipe.schema.RecipeSchema;
+
+public interface CrusherSchema {
+    RecipeKey<InputItem> INPUT = ItemComponents.INPUT.key("input");
+    RecipeKey<OutputItem> OUTPUT = ItemComponents.OUTPUT.key("output");
+    RecipeSchema SCHEMA = new RecipeSchema(OUTPUT,INPUT);
+}