+++ changelog.md
@@ -8,11 +8,14 @@
 - 添加了模组机器的配置文件
 
 ### v1.0.2
-- 添加了配方以及插件:速度/效率/加工的KubeJs支持
+- 添加了配方以及插件:速度/效率/加工的KubeJS支持
 
 ### v1.0.3
-- 为液态龙息以及液态又匿物质添加了放置粒子效果以及声音(可在模组对应的config文件配置)
+- 为液态龙息以及液态幽匿物质添加了放置粒子效果以及声音(可在模组对应的config文件配置)
 - 添加了龙之星发电机
+
+### v1.0.4
+- 添加了工业先锋所有支持json文件添加配方的KubeJS支持
 
 ## en_us
 ### v1.0.0
@@ -22,8 +25,11 @@
 - Added Config file for mod machines
 
 ### v1.0.2
-- Added KubeJs support for "recipe and Addon: speed/efficiency/processing"
+- Added KubeJS support for "recipe and Addon: speed/efficiency/processing"
 
 ### v1.0.3
 - Added particles and sound for liquid dragon breath and liquid sculk matter (can be configured in the corresponding config file of the mod)
 - Added Dragon Star generator
+
+### v1.0.4
+- Added KubeJS support for IndustrialForegoing's "All Supports Adding recipe to JSON Files" feature